Fixed test for razor view with generic model
VB CodeDom does not work on Mono/Linux (need vbnc)
CodeDomTypeReference has bug is parsing CLR type name
